build: create symlink to original glsl source file
Debugging runtime GLSL shader compilation is challenging because the
original GLSL source files are not copied to the binary directory when
compiled to SPIR-V as a custom build step. As a result, it's necessary
to manually copy GLSL source files to the binary directory to enable
runtime GLSL compilation which significantly degrades the debugging user
experience.

The reason GLSL source files had not been copied over is because they
live in the assets directory which is itself used as a symlink to avoid
copying large assets when building the project. Unfortunately you can't
create generated SPIR-V files into a symlink directory. Therefore, the
shaders directory has been moved out of the assets directory to allow
creating a symlink to the original GLSL source files so it can be used
to debug GLSL shader compilation issues at runtime if necessary. The use
of symbolic links, instead of copying files, was used to avoid potential
synchronization issues between the source and binary files.