fix: exception thrown when reading glsl files
This commit fixes a bug where reading a GLSL file would throw an
exception due the std::ios::failbit being set when reading from the
ifstream. This went unnoticed, in part, due to the difficulty with
debugging GLSL runtime shader compilation which was addressed in the
previous commit. However, this went unnoticed even before that because
exceptions were not enabled for ifstreams when reading GLSL files.

Upon further investigation, this relates to how newline characters are
handled on Windows when opening an ifstream in text mode. When tellg is
used after opening the file with std::ios::ate to get the size of the
file in bytes, it does not reflect the actual size of the file due to
automatic newline conversion which results in the file being read
incorrectly when used with std::ifstream::read which reads the file in
raw bytes.

To fix this, the GLSL file is opened in binary mode to avoid automatic
newline conversion. An alternative approach was considered that used
std::ostringstream to read the contents of an std::ifstream into a
string while handling newline conversions which would have been more
concise and idiomatic in C++. However, this approach was deemed
unnecessary because glslang already handles differences in newline
characters across platforms.