refactor: use spanstream to read GLSL/SPIR-V files
The current implementations to read SPIR-V and GLSL files endeavor to
avoid unnecessary copies by preallocating a fixed size buffer to stream
file contents into. When opening a file in text mode, this also requires
moving a string buffer into and out of an output stream which is error
prone and cumbersome.

Fortunately, C++23 added support for `std::spanstream` designed for this
exact scenario to enable streaming into a fixed-size buffer without
additional copies that unifies the implementations for reading text and
binary files.