feat: enable runtime SPIR-V optmization
This commit adds support for runtime SPIR-V optimization using the C++
interface for `spirv-opt` from the SPIR-V Tools library. For now, SPIR-V
optimizations are set to favor performance in release builds although
this has been implemented to trivially enable optimizations for size for
planned work to support mobile devices in the future.