feat: add Vulkan specialization constants support
The number of lights used in rendering is currently hardcoded in the
fragment shader which limits the types of scenes that are supported.
While this data can be passed in through a uniform buffer, this prevents
certain optimizations such as loop unrolling because the this value is
not known by the compiler when creating the shader module. Because this
value is known at pipeline creation time, a superior alternative is to
utilize specialized constants which allows the driver to override these
values during shader compilation.