test: add camera unit tests
This commit adds unit tests to verify camera position and orientation
after translations and rotations are applied. To simplify similar logic
with varying inputs, parameterized tests are used to generate tests
for different translation directions and rotation angles. Lastly, death
tests are added to verify that debug assertions are raised when created
with invalid arguments.