style: use .glsl file extension for shaders
This commit updates the naming convention for GLSL shaders to use .glsl
as the file extension as opposed to the glslang shader stage. While
convenient for auto stage detection in glslang, the use of different
extensions for unique shader stages results in inconsistent extensions
for GLSL files which is unnecessary when the shader stage simply be
included in the filename itself. Furthermore, this adds no additional
overhead because a shader compilation already knows what stage its
targeting during graphics pipeline creation.