build: simplify CMake presets
This commit implements several improvements aimed at simplifying CMake
preset configuration.

The first is the removal of clang-tidy presets which broke when the
project was ported to use C++20 modules due to a lack of support. The
preset remains hidden so it can be easily reinstated when clang tooling
is updated.

The second is to remove the "x64" and "x86" presets. It is incorrect to
separate "x64" and "ninja" because they both rely on the "architecture"
property and CMake presets do not support merging properties when
inherited outside of cache variables.

The last is merge the "ninja" and "default" presets since they are
always inherited together. While it would make sense to separate these
configurations if another generator is introduced, it is currently
unnecessary.