build: move compile and link options to presets
Despite using CMake presets to configure this project, many compile and
link options which depend on the target platform and build type are
defined in the root-level CMakeLists.txt file. This design is inflexible
when introducing configurations that rely on the same parameters (e.g.,
a debug configuration with or without AddressSanitizer enabled) which
becomes trivial when using CMakePresets.

To address this, compile and link options have been moved to CMake
presets. Unfortunately, adding CMAKE_CXX_FLAGS or CMAKE_EXE_LINKER_FLAGS
to a CMake preset will overwrite CMake's default values so to avoid this
compile and link options are instead defined as generic cache variables
which are then appended to CMAKE_CXX_FLAGS and CMAKE_EXE_LINKER_FLAGS in
CMakeLists.txt.