fix: resolve null buffer validation layers error
When updating material descriptor sets, `resize` instead of `reserve`
was erroneously used to set the initial vector's capacity which later
caused references descriptor buffer info to become invalidated when used
with `emplace_back`. Interestingly, this error does not reproduce in
release builds nor in debug builds with address sanitizer indicating
that the compiler was somehow optimizing around this. Fortunately, the
fix is fairly simple: just use `reserve`. Ideally, this method should
be reworked to avoid using `emplace_back` with references which can be
dangerous, but that is being saved for a larger refactor of the entire
glTF scene class.