fix: do not use infinity with --ffast-math
This commit fixes a clang warning when using infinity with --ffast-math
because it applies optimizations that assumes user code does not work
with infinity. In this case, it was used for infinite perspective which
was added in anticipation of supporting multiple glTF cameras but
currently isn't used. If/when support for this feature is added, it can
be reinstated using a dedicated sential value like FLOAT_MAX instead.