refactor: rework descriptor sets abstraction
This commit implements a number of minor improvements to the descriptor
sets abstraction.

This first is to rename `DescriptorSets` to `DescriptorPool` to better
represent the fact that is it both a descriptor pool and a set of
allocated descriptor sets. This naming convention leads to a more
natural API when accessing descriptor sets and their associated layout.
This naming convention also aligns with planned work to create a command
pool and associated allocated command buffers abstraction.

The second is to remove the default constructor for `DescriptorSets`
which precludes class invariants that guarantee the object is always in
a valid state. The reason it was added to deal with technical debt in
glTF scene loading which does not utilize initialization lists to ensure
objects are only initialized once. Unfortunately, fixing that is out of
scope for this commit. Instead, technical debt is isloated to glTF scene
loading by using std::optional to wrap that type with a TODO to address
this as previously described.

The last change here is aimed at general code cleanup with creating
global and material descriptor pools including adding and improving
assertions, comments, variable names, and avoiding magic constants.