refactor: remove returning transform view
The return value of `Swapchain::image_views` is a ranges::transform_view
over a `std::vector<vk::UniqueImageView>` that dereferences each image
view to return a `vk::ImageView`. The motivation behind this was that
a client is realistically only interested in a `vk::ImageView` and not
its vk::UniqueImageView counterpart. However, upon further review, this
unnecessarily complicates what should otherwise be a simple getter.
Furthermore, this diverges from how a similar type is handled in
`vktf::CommandPool`.