refactor: simplify sampler creation
When creating a sampler, the max LOD used for selecting mipmaps is the
maximum number of mipmap levels to utilize all mipmaps. This is a common
enough scenario that Vulkan provides a special value `VK_LOD_CLAMP_NONE`
that can be used to turn off clamping altogether. This helps simplify
the code by removing the need to load the texture from disk first before
creating any samplers which removes a dependency with task scheduling.