refactor: improve const correctness
This commit includes `const` correctness by utilizing pointer to const
in two places where it is was previously not thought possible to do so.
The first is light data which is copied each frame to be updated with
positions relative to its associated node world orientation. The second
is making a material pointer to const for primitives which only need
read access during rendering.