refactor: remove COLOR_0 vertex attribute support
This commit removes support for the optional COLOR_0 vertex attribute
which was hastily implemented with a suboptimal design that included
filling the vertex buffer with default values when undefined. Although
this vertex attribute was used by the current glTF model at one point,
its values were redundant (e.g., always zero) which were removed when
optimized through gltfpack. Therefore, because this feature is both
unused and adds complexity to this project it is being removed until a
later time when it can undergo a more rigorous design as part of dynamic
pipeline generation.