feat: create a mesh abstraction
This commit continues to whittle away at the massive amount of code in
`gltf_scene.cppm` by defining an abstraction for a mesh primitive which
consists of vertices, indices, and a non-owning material (managed by the
top-level scene). In addition to rendering, it also defines an API for
copying vertex and index data to device memory. Because this data is not
needed after copying to device memory, a `std::variant` is used (similar
to the approach found in `texture.cppm`) to represent vertex data as it
is transferred from host to device memory.

In addition to a mesh primitive, the new `mesh.cppm` module also exports
utility types for managing an index buffer which can be dynamic based on
its index type (i.e., `uint16_t` or `uint32_t`). The introduction of a
mesh primitive should naturally lead to a mesh abstraction also being
introduced, however, at this time it would simply consists of a `vector`
of primitives and is instead represented using a simple type alias.

When implementing the API to copy primitive data to device memory, it
was noted that the current design of returning host/device buffers pairs
is inflexible because it requires propagating staging buffers up several
function calls via `std::move` which is unwieldy. Therefore, the API was
simplified to receive a `vector` of staging buffers as an out parameter
which greatly simplifies logic in addition to making it more performant
by directly emplacing into the staging buffers vector. Although it's
considered best practice to use return values over output parameters,
the benefits of this approach justify the means. For consistenc, a
similar API in `texture.cppm` has been reworked to follow similar
calling conventions.

Lastly, out of necessity a new `material.cppm` module is introduced
which contains the base PBR metallic-roughness materials needed. This
is expected to undergo significant refactoring in a future commit where
an API is designed around constructing and user materials in a scene.