perf: optimize descriptor set bindings
During rendering, material descriptor sets are bound on a per-primitive
basis which is inefficient because materials are commonly shared across
mesh primitives. To fix this, rendering has been refactored so that
primitives sharing a common material are rendered together to reduce
descriptor set binding frequency as demonstrated in the following pseudo
code:

  foreach mesh by material
    bind material descriptor set
    foreach primitive in mesh
       push_constants := primitive.world_transform
       render primitive

One problem encountered with this implementation is that node transforms
are calculated dynamically when traversing the scene graph so rendering
primitives by material must be done in two passes: one to calculate the
world transform for each mesh primitive while grouping by material and
another to perform the actual rendering as described in the pseudocode
above.

This commit was inspired by the following NVIDIA blog post:
https://developer.nvidia.com/vulkan-shader-resource-binding