fix: build break in release builds
This commit fixes a build break caused by a warning (treated as an
error) for an unused parameter used only in a debug assertion. To ensure
errors such as this do not reoccur in the future, a task has been
recorded to look into getting a CI integration system working in GitHub
with appropriate pre-commit hooks.