feat: add staging primitive abstraction
Until now, a pattern for abstractions that manage device memory (e.g.,
buffers, images) has been to use std::variant to represent data as it
transfers from host to device memory. While convenient, it has two
serious disadvantages:
  1. It introduces error-prone APIs that depend on the object being in a
     particular state.
  2. Staging buffer management was decoupled from this abstraction which
     which makes multithreaded command buffer recording difficult.

To address these issues, a data-oriented approach is used where two
separate abstractions are introduced to represent data as it is
transferred from host-visible to device-local memory. By implementing
this in a purely transformational way, it allows for multithreaded
command buffer recording and potentially the beginning of some type of
job system.

This commit is a first attempt at implementing this pattern by adding a
staging primitive abstraction that represents vertex and index buffers
on host-visible memory which is later transformed to a primitive on
device-local memory.