feat: add host-visible buffer abstraction
This commit introduces a dedicated abstraction for host-visible buffers
which defines APIs for mapping, copying, and unmapping memory needed
in scenarios such creating staging buffers to transfer to device-local
memory and persistently mapped uniform buffers.

The previous approach was problematic because those APIs were defined on
a single buffer class which could be easily misused if the buffer wasn't
created with the correct memory property property flags needed to ensure
the buffer is created in host visible memory. This commit addresses this
defining those APIs on a dedicated type guaranteed to use the correct
memory property flags.

To avoid code duplication relating to buffer lifetime management, the
host-visible buffer class is derived from the base buffer class. While
this is sometimes discouraged due to the overhead associated with the
vtable, only the destructor is marked as virtual and therefore the
performance impact should be minimal. Furthermore, favoring composition
over inheritance would mean duplicating APIs in the base buffer class
which seems misguided.