feat: add graphics pipeline abstraction
This commit moves graphics pipeline creation out of the scene module
into a dedicated abstraction. In addition to making the code easier to
reason about, it also fixes technical debt associate with hardcoding the
offset for PushConstants::model_transform in Model::Render which was
previously required due to the PushConstants struct not being accessible
in the private module fragment of the scene module.