feat: improve numerical stability in glsl shaders
This commit implements a number of improvements to improve numerical
stability and code style.

The most notable change is the normal matrix is now constructed in the
fragment shader. Previously, this matrix was constructed in the vertex
shader and interpolated for each fragment which can result in floating
point precision issues where those basis vectors can become non-unit
length and lose orthogonality. Since each basis vector would need to be
renormalized, it makes more sense to simply pass the world-space normal
and tangent to the fragment shader so it can be constructed there which
also guarantees orthogonal basis vectors. This commit also fixes two
instances of a potential division by zero in pathological cases by
adding a small epsilon value to the divisor.

The next most important change here is the adoption of being explicit
about what coordinate space a vector resides in to avoid a common class
of bugs when working with shaders. This can be seen in several examples
such as world_position, world_normal, etc. To avoid confusion with a
view-space position, the fragment shader push constant has been renamed
to camera_world_position which is more clear about what it represents.
This also involved renaming Light to WorldLight to both represent the
coordinate space the light position is in and to align with the same
type used to copy data to the uniform buffer object in Scene::Update.
Lastly the convention texcoord_0 over texture_coordinates_0 is used to
align with the same conventions used in the glTF specification and this
project's glTF loading implementation.