build: rename graphics directory to engine
Historically the "graphics" directory was named after the gfx namespace
that was used for all files contained in this folder. However, a better
name for this directory is "engine" given the files in the directory are
used to create static library whose principal export is the vktf::Engine
class which is used by clients to load and render a scene.