feat: implement quaternion-based FPS camera
The current camera implementation is based on Euler angles which are
well known to suffer from a problem known as Gimbal lock where aligning
two rotational axes results in a loss of degree of freedom. Additionally
this camera implementation is inferior because it tries to ensure O(1)
access to view and projection transforms at the expensive of several CPU
intensive trigonometric functions to derive the camera orientation from
the view transform when performing rotations.

A superior alternative is to use quaternions which are the established
gold standard for camera implementations because they don't suffer from
Gimbal lock, have a lower memory footprint, and can be more efficiently
interpolated. This change also includes updated unit tests to account
for this fundamental change to the camera implementation.