fix: avoid referencing null gltf child nodes
This commit addresses a non-deterministic debug assertion that is
sometimes raised which is caused by a very subtle use-after-move bug
during glTF loading. The bug stems from the following code:

mutable_nodes
  | std::views::transform([&mutable_nodes](auto& key_value_pair) {
      auto& [cgltf_node, mutable_node] = key_value_pair;
      mutable_node->children = GetChildren(*cgltf_node, mutable_nodes);
      return std::pair{cgltf_node,
                  std::unique_ptr<const Node>(std::move(mutable_node))};
    })
  | std::ranges::to<std::unordered_map>();

which attempts to convert a node to a constant pointer after it has been
fully initialized. This unfortunately doesn't work because this requires
a std::move which invalidates the mutable node pointer which may still
need to be referenced by other mutable nodes which have not yet been
iterated over. The reason this occurred non-deterministically is because
the iteration order of an unordered_map is not guaranteed.

The fix for this requires creating nodes using three passes instead of
two:
  1. Create a nodes without establishing parent-child relationships
  2. Assign children pointers after all nodes have been created
  3. Convert to constant pointers after all nodes have been completely
     initialized.