refactor: add structs for each vertex attribute
This commit reworks the design for processing vertex attributes when
loading a glTF asset which previously used a type-alias for a vector of
floating point data to represent a vertex attribute in combination with
a sibling member property to represent the vertex attribute name. While
sufficient, this design is not extensible to support other related
vertex attribute properties such as Axis-Aligned Bounding Box (AABB)
information with the position attribute.

To address this vertex attributes have been refactored to use individual
structs for each vertex attribute by grouping vertex attribute data and
its associated glTF attribute name together. In addition to making the
API more type-safe and explicit, it also supports the addition of AABB
to the position attribute which will be needed for planned work to add
view-frustum culling.