feat: calculate mesh primitive bounding box
This commit adds support for calculating mesh primitive axis-aligned
bounding boxes (AABB) for future use in view-frustum culling. To achieve
this, it reads the min/max properties from the gltF mesh primitive
position attribute accessor which is required by the glTF specification.
It then marshalls this data to device-local mesh primitive abstraction
for use during rendering.

It's worth noting this change includes a minor refactoring to remove the
ToVec utility method. While this method was robust and type-safe over
glm::make_* APIs, it doesn't work well with subranges and refactoring it
to do so resulted in awkward user-experience by requiring an explicit
conversion to a std::span first which made the utility function less
useful.