refactor: avoid unnecessary material coupling
The implementation for a device-local mesh primitive stores a pointer to
a material which is unnecessary because only its descriptor set is used
during rendering. In addition to improving build times by removing this
module dependency, this should slightly improve performance by avoiding
an unnecessary branch in the core render loop which checked if a mesh
primitive is non-null before attempting to rendering a primitive which
is guaranteed to be true by staging primitive construction (which will
only contain a value if its associated material is non-null).