refactor: improve handling of empty data views
This commit adds explicit support for handling an empty data view by
first providing a default constructor which will initialize its data
pointer to nullptr and set the size to zero. It also improves error
handling in the constructor that accepts a pointer and size by verifying
that size must be zero if the data pointer is nullptr.