feat: implement AddKeyEventListener API
Keyboard event handling is limited to polling which complicates discrete
keyboard events detection (i.e., a single key press). Instead, a better
approach is to use event listeners that allow a caller to subscribe to
keyboard events.

This commit adds support for this scenario by implementing the
AddKeyEventListener API to the Window class which supports registering
multiple event listeners to invoke when a keyboard event is detected.